removed all integer(kind=), replaced with integer

+++ lib/tfmm3dlr/numthetafour.f
@@ -1,5 +1,5 @@
-	subroutine numthetafour(numtets,nlams)
-	integer *4 numtets(nlams),nlams 
+        subroutine numthetafour(numtets,nlams)
+        integer  numtets(nlams),nlams 
 c
 c This routine returns the number of Fourier modes needed in the
 c phi integral for each of the discrete lambda values given
