+++ src/cisurf_plottools.f90
@@ -5,12 +5,12 @@ implicit none
 
 !List of calling arguments
 character (len=100), intent(in) :: nombre
-integer ( kind = 8 ), intent(in) :: N
+integer, intent(in) :: N
 real ( kind = 8 ), intent(in) :: x(N),y(N)
 
 !List of local variables
 !character (len=100) nombre
-integer ( kind = 8 ) umio,count,flag
+integer umio,count,flag
 integer :: ierror
 !    nombre='fortran_plot'
     open(8, FILE=nombre,STATUS='REPLACE')
@@ -38,12 +38,12 @@ implicit none
 
 !List of calling arguments
 character (len=100), intent(in) :: nombre
-integer ( kind = 8 ), intent(in) :: N
+integer, intent(in) :: N
 real ( kind = 8 ), intent(in) :: x(N),y(N),z(N)
 
 !List of local variables
 !character (len=100) nombre
-integer ( kind = 8 ) umio,count,flag
+integer umio,count,flag
 integer :: ierror
 !    nombre='fortran_plot'
     open(8, FILE=nombre,STATUS='REPLACE')
@@ -73,12 +73,12 @@ implicit none
 
 !List of calling arguments
 character (len=100), intent(in) :: nombre
-integer ( kind = 8 ), intent(in) :: N,M
+integer, intent(in) :: N,M
 real ( kind = 8 ), intent(in) :: x(M,N),y(M,N),F(M,N)
 
 !List of local variables
 !character (len=100) nombre
-integer ( kind = 8 ) umio,count1,count2,flag
+integer umio,count1,count2,flag
 integer :: ierror
 !    nombre='fortran_plot'
     open(8, FILE=nombre,STATUS='REPLACE')
@@ -114,12 +114,12 @@ implicit none
 
 !List of calling arguments
 character (len=100), intent(in) :: nombre
-integer ( kind = 8 ), intent(in) :: N,M
+integer, intent(in) :: N,M
 real ( kind = 8 ), intent(in) :: F(M,N)
 
 !List of local variables
 !character (len=100) nombre
-integer ( kind = 8 ) umio,count1,count2,flag
+integer umio,count1,count2,flag
 integer :: ierror
 !    nombre='fortran_plot'
     open(8, FILE=nombre,STATUS='REPLACE')
@@ -144,11 +144,11 @@ implicit none
 
 !List of calling arguments
 character (len=100), intent(in) :: nombre
-integer ( kind = 8 ), intent(in) :: N,M
+integer, intent(in) :: N,M
 real ( kind = 8 ), intent(in) :: x(M,N),y(M,N),z(M,N),F(M,N)
 
 !List of local variables
-integer ( kind = 8 ) umio,count1,count2,flag
+integer umio,count1,count2,flag
 integer :: ierror
 !    nombre='fortran_plot'
     open(8, FILE=nombre,STATUS='REPLACE')
@@ -189,11 +189,11 @@ implicit none
 
 !List of calling arguments
 character (len=100), intent(in) :: nombre
-integer ( kind = 8 ), intent(in) :: num_box,n_pts
+integer, intent(in) :: num_box,n_pts
 real ( kind = 8 ), intent(in) :: W_boxes(num_box*4),Pts(3,n_pts)
 
 !List of local variables
-integer ( kind = 8 ) umio,count1,count2,flag
+integer umio,count1,count2,flag
 integer :: ierror
 !    nombre='fortran_plot'
     open(8, FILE=nombre,STATUS='REPLACE')
